+++ source/mtree_float_util.h
@@ -13,6 +13,6 @@ bool mtree_float_overlap_distance(mtree_float* first, mtree_float* second, float
 bool mtree_float_contains_distance(mtree_float* first, mtree_float* second, float distance);
 bool mtree_float_contained_distance(mtree_float* first, mtree_float* second, float distance);
 mtree_float* mtree_float_deep_copy(mtree_float* source);
-int get_float_distance(int size, mtree_float* entries[size], int distances[size][size], int i, int j);
+float get_float_distance(int size, mtree_float* entries[size], float distances[size][size], int i, int j);
 
 #endif
